move edit storage engine
