+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/di/dependencies/EngineDependencies.kt
@@ -1,6 +1,6 @@
 package com.github.klee0kai.thekey.app.di.dependencies
 
-import com.github.klee0kai.thekey.app.engine.findstorage.EditStorageEngine
+import com.github.klee0kai.thekey.app.engine.editstorage.EditStorageEngine
 import com.github.klee0kai.thekey.app.engine.findstorage.FindStorageEngine
 import com.github.klee0kai.thekey.app.engine.storage.CryptStorage
 import com.github.klee0kai.thekey.app.engine.storage.CryptStorageSuspended
